consolidate type definitions
